Gzip the man page when installing
When the reference client is installed, the gzipped man page needs
to be copied to the appropriate system location. The gzipped man
page is currently committed in the repository. A better approach is
to gzip the man page at installation time. This removes the gzipped
man page and applies the better approach.