Add instance class Humanoid
Humanoid will be the base of all character objects in OpenBlox.
Developers will never be required to use the Humanoid class to
create characters, but the goal is to make it easy to create
character objects.

In the current plan, the Humanoid class will tie in with the
optional CoreUI. With the current plan, if the following:

game.Players.CurrentPlayer.Character and
game.Players.CurrentPlayer.Character:FindFirstChildOfClass("Humanoid")

is true, then CoreUI will show a basic health bar if enabled.