Add instance class CoreGui
The CoreGui service holds persistent UI elements that should be
presented to the user regardless of whether or not there is a Player
object in the current game. This makes is very useful for showing
server debug information or information for server administrators,
or perhaps a pre-connection menu or loading screen for clients
connecting to a server.