Fix AssetLocator
AssetLocator was returning an uninitalized
shared_ptr<AssetLocator>. This caused some issues with shared_ptr
deinitialization, which eventually caused a segmentation fault.

ContentProvider now has the AssetLoaded event, which is fired
whenever AssetLocator loads an asset.

TaskScheduler has been altered so that it can run as a FIFO queue,
if that's desired.