Update OBEngine to use new secondary task scheduler
The secondary TaskScheduler has been set up to run as a FIFO
queue. This is more of an efficiency change than anything else,
because we always pass things to the secondary scheduler with an
execution time of 0. Sorts don't alter the order at all, so we're
really just skipping sorts where they wouldn't do anything.