Fix Events
Events had several issues, one being that VarWrapper wasn't
properly handling deinitialization. Deinitialization is a big
pain in the ass, because 'wrapped' can either be a C++ class or
a random C pointer, and we don't have any common classes. This
leads to casting of 'wrapped' to whatever type it actually is,
before calling 'delete' or 'free' on it.

In addition, there was an issue with Lua event connections failing
leading to event connections only running once if a Lua error
occured. This has been fixed in Event.cpp. There was a similar
error that often occured that led to a segmentation fault after
lua_close was called on the Lua state returned. For now, we've
'fixed' this by just commenting out the calls to 'lua_close'. This
is not permanent.