Add properties to Workspace, change type of Gravity property
Gravity property is now a Vector3, allowing directional gravity.

FallenPartsDestroyHeight is set to -1000 by default, and when
DestroyFallenParts is true, parts that fall below that will be
Destroy()'d. This is not true of Anchored parts.

DestroyFallenParts defaults to true, but this may change in the
future.

Gravity is now incompatible with ROBLOX, but that shouldn't be an
issue. Gravity on ROBLOX is a double, but luckily it can only be
set by ROBLOX Studio, so we shouldn't have to worry about it. If
it comes to it, when we implement ROBLOX place loading we can just
convert ROBLOX's Gravity to be Vector3.new(0, robloxGravity, 0)